fix: resolve new Qodana findings in codegeneration and push modules
- CA1034: suppress nested V1/V2 records in CodeGenerationConfigResult
  (intentional discriminated union pattern; private ctor prevents external subclassing)
- UnusedMember.Global: suppress Version property in CodeGenerationConfigBase
  (needed for JSON round-trip even though not accessed in code after deserialization)
- CollectionNeverUpdated.Local: add WithCustomConfiguration() to CodeGenerationContextBuilder
  (completes the builder API; the _customConfigurations list is already iterated in Build())
- UseCollectionExpression: replace new[] { ... } with [...] in ConfigMappingTests
- UseObjectOrCollectionInitializer: use Attributes = { [...] } initializer syntax
  in WebresourcesProcessorTests for WebResource and SolutionComponent
- S104 + S4136: extract V1 CodeGenerationConfig overloads from MetadataService.cs
  into MetadataService.Legacy.cs (partial class)
  - MetadataService.cs: 1065 → 688 lines (fixes S104 file-too-long)
  - S4136 overload-adjacency no longer fires (overloads are in separate files)
- S4136 in IMetadataService: reorder interface so each V1 overload is immediately
  adjacent to its V2 counterpart; remove #region Legacy V1 support separator
