+++ src/modules/dgt.power.codegeneration/Base/CodeGenerationConfigBase.cs
@@ -4,6 +4,7 @@
 // ReSharper disable UnusedAutoPropertyAccessor.Global
 // ReSharper disable AutoPropertyCanBeMadeGetOnly.Global
 // ReSharper disable CollectionNeverUpdated.Global
+// ReSharper disable UnusedMember.Global
 
 namespace dgt.power.codegeneration.Base;
